fix: refactor http (take 2)
