+++ lib/util/http/errors.ts
@@ -1,0 +1,2 @@
+// required for zod type safety with `Result.wrap`
+export class EmptyResultError extends Error {}
