+++ lib/util/http/jira.ts
@@ -1,5 +1,5 @@
+import { HttpBase } from './http';
 import type { HttpOptions } from './types';
-import { Http } from '.';
 
 let baseUrl: string;
 
@@ -7,7 +7,7 @@ export function setBaseUrl(url: string): void {
   baseUrl = url;
 }
 
-export class JiraHttp extends Http {
+export class JiraHttp extends HttpBase {
   protected override get baseUrl(): string | undefined {
     return baseUrl;
   }
